Improve performance of `:keys` validation interpreter
Previously, the validation performance for `:keys` was not a concern
because it was done simultaneously while navigating a schema
i.e. while navigating through keys in incoming data we just checked
whether they exist in schema. Now, navigation and validation are done
separately, and we have to perform validation as efficiently as
possible to avoid regressions.